backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
google-generativeai
python-dotenv
httpx
python-multipart
